Inline the old ObjectLocationPrefixTransfer trait
